requirements.txt
torch
torchvision
numpy
matplotlib
scipy
open-cv
torchsummary
tensorboard
pillow
tqdm